Fix broken local storage algorithm
#### Here's why:
Seeing some restructuring were made on the index.html file, come classes and ids have changed.
I figured that it is actually more helpful to peg the event listening on the button instead. This increase the quality of data stored without having to use up too much space or modify values as the user only submits on the form upon assurance of his entries.